+++ pkg/kubernetes/report_client_test.go
@@ -65,8 +65,8 @@ func Test_ResultClient_FetchPolicyResults(t *testing.T) {
 	mapper := NewMapper(k8sCMClient)
 
 	client := kubernetes.NewPolicyResultClient(
-		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now()),
-		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now()),
+		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5),
+		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5),
 	)
 
 	fakeAdapter.policies = append(fakeAdapter.policies, unstructured.Unstructured{Object: policyMap})
@@ -92,8 +92,8 @@ func Test_ResultClient_FetchPolicyResultsPolicyReportError(t *testing.T) {
 	mapper := NewMapper(k8sCMClient)
 
 	client := kubernetes.NewPolicyResultClient(
-		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now()),
-		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now()),
+		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5),
+		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5),
 	)
 
 	_, err := client.FetchPolicyResults()
@@ -112,8 +112,8 @@ func Test_ResultClient_FetchPolicyResultsClusterPolicyReportError(t *testing.T) 
 	mapper := NewMapper(k8sCMClient)
 
 	client := kubernetes.NewPolicyResultClient(
-		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now()),
-		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now()),
+		kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5),
+		kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5),
 	)
 
 	_, err := client.FetchPolicyResults()
@@ -129,8 +129,8 @@ func Test_ResultClient_RegisterPolicyResultWatcher(t *testing.T) {
 
 	mapper := NewMapper(k8sCMClient)
 
-	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now())
-	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now())
+	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5)
+	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5)
 
 	client := kubernetes.NewPolicyResultClient(pClient, cpClient)
 
@@ -166,8 +166,8 @@ func Test_ResultClient_SkipReportsWithoutResults(t *testing.T) {
 
 	mapper := NewMapper(k8sCMClient)
 
-	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now())
-	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now())
+	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5)
+	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5)
 
 	client := kubernetes.NewPolicyResultClient(pClient, cpClient)
 
@@ -239,8 +239,8 @@ func Test_ResultClient_SkipReportsCleanUpEvents(t *testing.T) {
 
 	mapper := NewMapper(k8sCMClient)
 
-	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now())
-	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now())
+	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5)
+	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5)
 
 	client := kubernetes.NewPolicyResultClient(pClient, cpClient)
 
@@ -312,8 +312,8 @@ func Test_ResultClient_SkipReportsReconnectEvents(t *testing.T) {
 
 	mapper := NewMapper(k8sCMClient)
 
-	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now())
-	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now())
+	pClient := kubernetes.NewPolicyReportClient(fakeAdapter, report.NewPolicyReportStore(), mapper, time.Now(), 5)
+	cpClient := kubernetes.NewClusterPolicyReportClient(fakeAdapter, report.NewClusterPolicyReportStore(), mapper, time.Now(), 5)
 
 	client := kubernetes.NewPolicyResultClient(pClient, cpClient)